added liquidbase

+++ src/main/java/com/phoneshop/model/response/ModelResponse.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class ModelResponse {
 
-	private Integer modelId;
-	private Integer brandId;
+	private Long modelId;
+	private Long brandId;
 	private String name;
 	private Double version;
