+++ src/main/java/com/phoneshop/repository/ModelRepository.java
@@ -3,6 +3,6 @@ package com.phoneshop.repository;
 import com.phoneshop.model.entity.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModelRepository extends JpaRepository<Model, Integer>{
+public interface ModelRepository extends JpaRepository<Model, Long>{
 
 }
