+++ src/main/java/com/phoneshop/service/ModelService.java
@@ -2,7 +2,6 @@ package com.phoneshop.service;
 
 import java.util.List;
 
-import com.phoneshop.dto.ModelDTO;
 import com.phoneshop.model.entity.Model;
 import com.phoneshop.model.response.ModelResponse;
 
@@ -10,7 +9,7 @@ public interface ModelService {
 
 	Model save(Model model);
 
-	ModelResponse getModelByModelId(Integer id);
+	ModelResponse getModelByModelId(Long id);
 	
 	List<Model> getAllModels();
